trackey/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
pillow==10.2.0
djangorestframework_simplejwt==5.3.1
django_cors_headers==4.3.1